chore: moves files to new directories
