+++ src/components/forms/OwnerResponseForm.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { API_BASE_URL } from '../config';
+import { API_BASE_URL } from '../../config';
 
 export const OwnerResponseForm = (props) => {
   const [ownerResponse, setOwnerResponse] = useState('');
