+++ src/components/plates/MyPlate.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import PagesNav from './PagesNav';
+import PagesNav from '../PagesNav';
 import UnclaimPlate from './UnclaimPlate';
-import Plate from './common/Plate';
-import { usePlate } from '../hooks/use-plate';
-import '../styles/plates/single-plate.css';
-import ReviewList from './common/ReviewList';
+import Plate from '../common/Plate';
+import { usePlate } from '../../hooks/use-plate';
+import '../../styles/plates/single-plate.css';
+import ReviewList from '../common/ReviewList';
 
 export const MyPlate = () => {
   const { plate, reviews } = usePlate();
