+++ src/components/plates/UnclaimPlate.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { API_BASE_URL } from '../config';
+import { API_BASE_URL } from '../../config';
 
 const UnclaimPlate = () => {
   const [unclaimMessage, setUnclaimMessage] = useState('');
